+++ bin/www
@@ -17,6 +17,8 @@ var app = main.setupApp();
 var port = normalizePort(process.env.PORT || '3000');
 app.set('port', port);
 
+console.log("Running on port " + port);
+
 /**
  * Create HTTP server.
  */
